paas2/login/requirements.txt
django==1.8.19
django-braces==1.13.0
dj-static==0.0.6
pycrypto==2.6.1
requests==2.21.0
pymysql==0.6.7
gunicorn==19.9.0
uWSGI==2.0.13.1
xlrd==1.0.0
xlwt==1.1.2
gevent==1.1.2
pytz==2016.6.1
python-dateutil==2.6.0
django_oauth_toolkit==0.12.0
django-decorator-include==1.3
cachetools==3.1.1

# for config
django-environ==0.4.5
# prometheus metrics
django-prometheus==1.0.15
